all go version should pass, added pre17 and 17 for replaceNetFunc to handle context problem
