+++ naming/dns_resolver_test.go
@@ -1,7 +1,6 @@
 package naming
 
 import (
-	"context"
 	"fmt"
 	"net"
 	"reflect"
@@ -181,32 +180,6 @@ var srvLookupTbl = map[string][]*net.SRV{
 	"_grpclb._tcp.srv.ipv4.multi.fake":  {newSRVRR("ipv4.multi.fake", 1234)},
 	"_grpclb._tcp.srv.ipv6.single.fake": {newSRVRR("ipv6.single.fake", 1234)},
 	"_grpclb._tcp.srv.ipv6.multi.fake":  {newSRVRR("ipv6.multi.fake", 1234)},
-}
-
-func replaceNetFunc() func() {
-	oldLookupHost := lookupHost
-	oldLookupSRV := lookupSRV
-	addrToResolve = fakeAddrToResolve
-	addrResolved = fakeAddrResolved
-	lookupHost = func(ctx context.Context, host string) ([]string, error) {
-		if addrs, ok := hostLookupTbl[host]; ok {
-			return addrs, nil
-		}
-		return nil, fmt.Errorf("failed to lookup host:%s resolution in hostLookupTbl", host)
-	}
-	lookupSRV = func(ctx context.Context, service, proto, name string) (string, []*net.SRV, error) {
-		cname := "_" + service + "._" + proto + "." + name
-		if srvs, ok := srvLookupTbl[cname]; ok {
-			return cname, srvs, nil
-		}
-		return "", nil, fmt.Errorf("failed to lookup srv record for %s in srvLookupTbl", cname)
-	}
-	return func() {
-		lookupHost = oldLookupHost
-		lookupSRV = oldLookupSRV
-		addrToResolve = realAddrToResolve
-		addrResolved = realAddrResolved
-	}
 }
 
 func updatesToSlice(updates []*Update) []Update {
